bootstrap now creates bin/reposervice-env
The goal here is to separate out the environment setup functionality
from the config file substitution and conditional moves.  The
'bootstrap' script, soon to be renamed 'bin/reposervice-config',
generates the 'bin/reposervice-env' file of bash commands to set up the
environment.  A future bootstap script will use that file rather than
the existing process for setting up the environment.