Support JP2 in book and large image solution pack
Note: This is a reworking of the previous native JPEG2000 support
in the large image solution pack.  We no longer create and store
the TIFF derivative if the original OBJ is a JPEG2000.  We will
create a temporary derivative TIFF of a JP2 to perform the OCR
because tesseract does not natively support JP2 files.

Also note: with this commit, Jasper is a prerequisite and ImageMagick
must be configured to use Jasper for handling JP2 files.